move comps to ui folder
