+++ app/javascript/components/items-list/items-list-component.jsx
@@ -1,7 +1,7 @@
 import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
 
-import Icon from 'components/icon';
+import Icon from 'components/ui/icon';
 import arrowIcon from 'assets/icons/flechita.svg';
 import './items-list-styles.scss';
