+++ app/javascript/components/map/components/mini-legend/mini-legend-component.jsx
@@ -2,9 +2,9 @@ import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
 import { isTouch } from 'utils/browser';
 
-import Button from 'components/button';
-import Icon from 'components/icon';
-import Tip from 'components/tip';
+import Button from 'components/ui/button';
+import Icon from 'components/ui/icon';
+import Tip from 'components/ui/tip';
 
 import linkIcon from 'assets/icons/link.svg';
 import './mini-legend-styles.scss';
