+++ app/javascript/components/ui/dropdown/components/item.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import Icon from 'components/icon';
-import Button from 'components/button';
+import Icon from 'components/ui/icon';
+import Button from 'components/ui/button';
 
 import arrowDownIcon from 'assets/icons/arrow-down.svg';
 import infoIcon from 'assets/icons/info.svg';
