+++ app/javascript/components/ui/search/search-component.jsx
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import Icon from 'components/icon';
-import Button from 'components/button';
+import Icon from 'components/ui/icon';
+import Button from 'components/ui/button';
 import debounce from 'lodash/debounce';
 
 import searchIcon from 'assets/icons/search.svg';
