+++ app/javascript/pages/country/analytics.js
@@ -1,6 +1,6 @@
-import shareActions from 'components/share/share-actions';
-import modalActions from 'components/modals/meta/modal-meta-actions';
-import buttonActions from 'components/button/button-actions';
+import shareActions from 'components/modals/share/share-actions';
+import modalActions from 'components/modals/meta/meta-actions';
+import buttonActions from 'components/ui/button/button-actions';
 import { COUNTRY } from 'pages/country/router';
 import isEqual from 'lodash/isEqual';
 import omit from 'lodash/omit';
