+++ app/javascript/pages/country/footer/footer-component.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Button from 'components/button';
+import Button from 'components/ui/button';
 
 import './footer-styles.scss';
